Added Compile/Clean/Init command, module stubs
A module is analogue to a Package in Swift, it represents a compilation
unit. A module with sources that must be compiled a SrcModule can have
dependencies of various type, e.g. a GitModule that download a javabuild
package from a git repository, compiles it a places the outpur in
.dependencies, ready to be added to the classpath of the current
project.

Just a few ideas, everything could change.